Large refactor attmept
