+++ nvim/lua/plugins/completions.lua
@@ -1,0 +1,130 @@
+return {
+    "hrsh7th/nvim-cmp",
+  "hrsh7th/cmp-nvim-lsp",
+  "onsails/lspkind-nvim",
+  {
+    'L3MON4D3/LuaSnip',
+    config = function()
+      local cmp = require("cmp")
+      local luasnip = require('luasnip')
+
+      -- Load snippets from vscode
+      require("luasnip.loaders.from_vscode").lazy_load()
+
+      -- Setup cmp
+      cmp.setup({
+        snippet = {
+          expand = function(args)
+            require("luasnip").lsp_expand(args.body)
+          end,
+        },
+        window = {
+          completion = cmp.config.window.bordered(),
+          documentation = cmp.config.window.bordered(),
+        },
+        mapping = cmp.mapping.preset.insert({
+          ['<C-b>'] = cmp.mapping.scroll_docs(-4),
+          ['<C-f>'] = cmp.mapping.scroll_docs(4),
+          ['<C-o>'] = cmp.mapping.complete(),
+          ['<C-i>'] = cmp.mapping.abort(),
+          ['<CR>'] = cmp.mapping.confirm({ select = true }),
+        }),
+        formatting = {
+          fields = { 'menu', 'abbr', 'kind' },
+          format = function(entry, item)
+            local menu_icon = {
+              nvim_lsp = 'λ',
+              luasnip = '⋗',
+              buffer = 'Ω',
+              path = '🖫',
+            }
+            item.menu = menu_icon[entry.source.name]
+            return item
+          end,
+        },
+        sources = cmp.config.sources({
+          { name = 'nvim_lsp' },
+          { name = 'luasnip' },
+        }, {
+          { name = 'buffer' },
+        }),
+        ['<C-b>'] = cmp.mapping(function(fallback)
+          if luasnip.jumpable(-1) then
+            luasnip.jump(-1)
+          else
+            fallback()
+          end
+        end, { 'i', 's' }),
+
+        ['<Tab>'] = cmp.mapping(function(fallback)
+          local col = vim.fn.col('.') - 1
+
+          if cmp.visible() then
+            cmp.select_next_item(select_opts)
+          elseif col == 0 or vim.fn.getline('.'):sub(col, col):match('%s') then
+            fallback()
+          else
+            cmp.complete()
+          end
+        end, { 'i', 's' }),
+
+        ['<S-Tab>'] = cmp.mapping(function(fallback)
+          if cmp.visible() then
+            cmp.select_prev_item(select_opts)
+          else
+            fallback()
+          end
+        end, { 'i', 's' }),
+      })
+
+      -- Define custom signs for diagnostics
+      local sign = function(opts)
+        vim.fn.sign_define(opts.name, {
+          texthl = opts.name,
+          text = opts.text,
+          numhl = ''
+        })
+      end
+
+      sign({ name = 'DiagnosticSignError', text = '✘' })
+      sign({ name = 'DiagnosticSignWarn', text = '▲' })
+      sign({ name = 'DiagnosticSignHint', text = '' })
+      sign({ name = 'DiagnosticSignInfo', text = '' })
+
+      vim.diagnostic.config({
+        virtual_text = false,
+        severity_sort = true,
+        float = {
+          border = 'rounded',
+          source = 'always',
+        },
+      })
+
+      -- Set up LSP hover and signature help with rounded borders
+      vim.lsp.handlers['textDocument/hover'] = vim.lsp.with(
+        vim.lsp.handlers.hover,
+        { border = 'rounded' }
+      )
+
+      vim.lsp.handlers['textDocument/signatureHelp'] = vim.lsp.with(
+        vim.lsp.handlers.signature_help,
+        { border = 'rounded' }
+      )
+    end
+  },
+
+  -- Setup vim-dadbod for SQL files
+  {
+    'kristijanhusak/vim-dadbod-completion',
+    ft = { "sql" },
+    config = function()
+      local cmp = require("cmp")
+      cmp.setup.filetype({ "sql" }, {
+        sources = {
+          { name = "vim-dadbod-completion" },
+          { name = "buffer" },
+        },
+      })
+    end
+  }
+}
