+++ nvim/lua/plugins/leap.lua
@@ -1,0 +1,6 @@
+return {
+  "ggandor/leap.nvim",
+  config = function()
+    require('leap').add_default_mappings()
+  end
+}
