+++ nvim/lua/plugins/neogit.lua
@@ -1,0 +1,7 @@
+return {
+  "TimUntersberger/neogit",
+  dependencies = { "nvim-lua/plenary.nvim" },
+  config = function()
+    require('neogit').setup {}
+  end
+}
