+++ nvim/lua/plugins/nvim-tree.lua
