+++ nvim/lua/plugins/prettier.lua
@@ -1,0 +1,24 @@
+return {
+  "jose-elias-alvarez/null-ls.nvim",
+  config = function()
+    local prettier = require("prettier")
+
+    prettier.setup({
+      bin = 'prettier',
+      filetypes = {
+        "css",
+        "graphql",
+        "html",
+        "javascript",
+        "javascriptreact",
+        "json",
+        "less",
+        "markdown",
+        "scss",
+        "typescript",
+        "typescriptreact",
+        "yaml",
+      },
+    })
+  end
+}
