+++ nvim/lua/plugins/telescope-file-browser.lua
@@ -1,0 +1,114 @@
+return {
+  -- Telescope Plugin
+  {
+    "nvim-telescope/telescope.nvim",
+    dependencies = { "nvim-lua/plenary.nvim", "nvim-telescope/telescope-file-browser.nvim" }, -- Telescope file browser extension
+    config = function()
+      local actions = require("telescope.actions")
+
+      require("telescope").setup({
+        pickers = {
+          find_files = {
+            find_command = { "rg", "--files", "--hidden", "-g", "!.git" },
+          },
+        },
+
+        defaults = {
+          file_sorter = require("telescope.sorters").get_fzy_sorter,
+          color_devicons = true,
+          file_previewer = require("telescope.previewers").vim_buffer_cat.new,
+          grep_previewer = require("telescope.previewers").vim_buffer_vimgrep.new,
+          qflist_previewer = require("telescope.previewers").vim_buffer_qflist.new,
+          path_display = { "smart" },
+
+          sorting_strategy = "ascending",
+          layout_strategy = "horizontal",
+          border = true,
+          layout_config = {
+            horizontal = {
+              prompt_position = "top",
+            },
+          },
+        },
+
+        extensions = {
+          fzy_native = {
+            override_generic_sorter = false,
+            override_file_sorter = true,
+          },
+          file_browser = {
+            hidden = { file_browser = true, folder_browser = true },
+            no_ignore = true,
+            hijack_netrw = true,
+            mappings = {
+              ["i"] = {
+                ["<C-n>"] = actions.cycle_history_next,
+                ["<C-p>"] = actions.cycle_history_prev,
+
+                ["<C-j>"] = actions.move_selection_next,
+                ["<C-k>"] = actions.move_selection_previous,
+
+                ["<C-c>"] = actions.close,
+
+                ["<Down>"] = actions.move_selection_next,
+                ["<Up>"] = actions.move_selection_previous,
+
+                ["<CR>"] = actions.select_default,
+                ["<C-x>"] = actions.select_horizontal,
+                ["<C-v>"] = actions.select_vertical,
+                ["<C-t>"] = actions.select_tab,
+
+                ["<C-u>"] = actions.preview_scrolling_up,
+                ["<C-d>"] = actions.preview_scrolling_down,
+
+                ["<PageUp>"] = actions.results_scrolling_up,
+                ["<PageDown>"] = actions.results_scrolling_down,
+
+                ["<Tab>"] = actions.toggle_selection + actions.move_selection_worse,
+                ["<S-Tab>"] = actions.toggle_selection + actions.move_selection_better,
+                ["<C-q>"] = actions.send_to_qflist + actions.open_qflist,
+                ["<M-q>"] = actions.send_selected_to_qflist + actions.open_qflist,
+                ["<C-l>"] = actions.complete_tag,
+                ["<C-_>"] = actions.which_key,
+              },
+              ["n"] = {
+                ["<esc>"] = actions.close,
+                ["<CR>"] = actions.select_default,
+                ["<C-x>"] = actions.select_horizontal,
+                ["<C-v>"] = actions.select_vertical,
+                ["<C-t>"] = actions.select_tab,
+
+                ["<Tab>"] = actions.toggle_selection + actions.move_selection_worse,
+                ["<S-Tab>"] = actions.toggle_selection + actions.move_selection_better,
+                ["<C-q>"] = actions.send_to_qflist + actions.open_qflist,
+                ["<M-q>"] = actions.send_selected_to_qflist + actions.open_qflist,
+
+                ["j"] = actions.move_selection_next,
+                ["k"] = actions.move_selection_previous,
+                ["H"] = actions.move_to_top,
+                ["M"] = actions.move_to_middle,
+                ["L"] = actions.move_to_bottom,
+
+                ["<Down>"] = actions.move_selection_next,
+                ["<Up>"] = actions.move_selection_previous,
+                ["gg"] = actions.move_to_top,
+                ["G"] = actions.move_to_bottom,
+
+                ["<C-u>"] = actions.preview_scrolling_up,
+                ["<C-d>"] = actions.preview_scrolling_down,
+
+                ["<PageUp>"] = actions.results_scrolling_up,
+                ["<PageDown>"] = actions.results_scrolling_down,
+
+                ["?"] = actions.which_key,
+              },
+            },
+          },
+        },
+      })
+
+      -- Load the telescope-file-browser extension
+      require("telescope").load_extension("file_browser")
+    end
+  }
+}
