--- nvim/lua/scottykaye/config/plugins/fugitive.lua
@@ -1,1 +1,0 @@
-vim.keymap.set("n", "<leader>gs", vim.cmd.Git);
