--- nvim/lua/scottykaye/config/plugins/leap.lua
@@ -1,1 +1,0 @@
-require('leap').add_default_mappings()
