--- nvim/lua/scottykaye/config/plugins/mason-lspconfig.lua
@@ -1,3 +1,0 @@
-require("mason-lspconfig").setup({
-  ensured_installed = { "lua_ls", "solargraph", "ts_ls", "biome", "eslint", "tailwindcss", "phpactor", "php-cs-fixer" },
-})
