--- nvim/lua/scottykaye/config/plugins/mason.lua
@@ -1,2 +1,0 @@
-require("mason").setup()
-vim.keymap.set("n", "<leader>cm", ":Mason<CR>", { noremap = true, silent = true })
