--- nvim/lua/scottykaye/config/plugins/neogit.lua
@@ -1,1 +1,0 @@
-require('neogit').setup {}
