--- nvim/lua/scottykaye/config/plugins/nvim-autopairs.lua
@@ -1,3 +1,0 @@
-require("nvim-autopairs").setup {
-  check_ts = true,
-}
