--- nvim/lua/scottykaye/config/plugins/scrollbar.lua
@@ -1,1 +1,0 @@
-require("scrollbar").setup()
