--- nvim/lua/scottykaye/config/plugins/telescope-file-browser.lua
@@ -1,114 +1,0 @@
--- You don't need to set any of these options.
--- IMPORTANT!: this is only a showcase of how you can set default options!
-local actions = require("telescope.actions")
-
-require("telescope").setup({
-  pickers = {
-    find_files = {
-      find_command = { "rg", "--files", "--hidden", "-g", "!.git" },
-    },
-  },
-
-  defaults = {
-    file_sorter = require("telescope.sorters").get_fzy_sorter,
-    color_devicons = true,
-    file_previewer = require("telescope.previewers").vim_buffer_cat.new,
-    grep_previewer = require("telescope.previewers").vim_buffer_vimgrep.new,
-    qflist_previewer = require("telescope.previewers").vim_buffer_qflist.new,
-    path_display = { "smart" },
-
-    sorting_strategy = "ascending",
-    layout_strategy = "horizontal",
-    border = true,
-    layout_config = {
-      -- Search bar to the top
-      horizontal = {
-        prompt_position = "top",
-      },
-    },
-  },
-
-  extensions = {
-    fzy_native = {
-      override_generic_sorter = false,
-      override_file_sorter = true,
-    },
-    file_browser = {
-      -- ivy is at the bottom file browser kinda cool but not right now!
-      -- theme = "ivy",
-      -- theme = "dropdown",
-      hidden = { file_browser = true, folder_browser = true },
-      no_ignore = true,
-      -- disables netrw and use telescope-file-browser in its place
-      hijack_netrw = true,
-      mappings = {
-        ["i"] = {
-          ["<C-n>"] = actions.cycle_history_next,
-          ["<C-p>"] = actions.cycle_history_prev,
-
-          ["<C-j>"] = actions.move_selection_next,
-          ["<C-k>"] = actions.move_selection_previous,
-
-          ["<C-c>"] = actions.close,
-
-          ["<Down>"] = actions.move_selection_next,
-          ["<Up>"] = actions.move_selection_previous,
-
-          ["<CR>"] = actions.select_default,
-          ["<C-x>"] = actions.select_horizontal,
-          ["<C-v>"] = actions.select_vertical,
-          ["<C-t>"] = actions.select_tab,
-
-          ["<C-u>"] = actions.preview_scrolling_up,
-          ["<C-d>"] = actions.preview_scrolling_down,
-
-          ["<PageUp>"] = actions.results_scrolling_up,
-          ["<PageDown>"] = actions.results_scrolling_down,
-
-          ["<Tab>"] = actions.toggle_selection + actions.move_selection_worse,
-          ["<S-Tab>"] = actions.toggle_selection + actions.move_selection_better,
-          ["<C-q>"] = actions.send_to_qflist + actions.open_qflist,
-          ["<M-q>"] = actions.send_selected_to_qflist + actions.open_qflist,
-          ["<C-l>"] = actions.complete_tag,
-          ["<C-_>"] = actions.which_key, -- keys from pressing <C-/>
-        },
-        ["n"] = {
-          ["<esc>"] = actions.close,
-          ["<CR>"] = actions.select_default,
-          ["<C-x>"] = actions.select_horizontal,
-          ["<C-v>"] = actions.select_vertical,
-          ["<C-t>"] = actions.select_tab,
-
-          ["<Tab>"] = actions.toggle_selection + actions.move_selection_worse,
-          ["<S-Tab>"] = actions.toggle_selection + actions.move_selection_better,
-          ["<C-q>"] = actions.send_to_qflist + actions.open_qflist,
-          ["<M-q>"] = actions.send_selected_to_qflist + actions.open_qflist,
-
-          ["j"] = actions.move_selection_next,
-          ["k"] = actions.move_selection_previous,
-          ["H"] = actions.move_to_top,
-          ["M"] = actions.move_to_middle,
-          ["L"] = actions.move_to_bottom,
-
-          ["<Down>"] = actions.move_selection_next,
-          ["<Up>"] = actions.move_selection_previous,
-          ["gg"] = actions.move_to_top,
-          ["G"] = actions.move_to_bottom,
-
-          ["<C-u>"] = actions.preview_scrolling_up,
-          ["<C-d>"] = actions.preview_scrolling_down,
-
-          ["<PageUp>"] = actions.results_scrolling_up,
-          ["<PageDown>"] = actions.results_scrolling_down,
-
-          ["?"] = actions.which_key,
-        },
-      },
-    },
-  },
-})
-
--- require('telescope').load_extension('fzy_native')
--- To get telescope-file-browser loaded and working with telescope,
--- you need to call load_extension, somewhere after setup function:
-require("telescope").load_extension("file_browser")
