--- nvim/lua/scottykaye/config/plugins/treesitter.lua
@@ -1,56 +1,0 @@
-require 'nvim-treesitter.configs'.setup {
-  -- A list of parser names, or "all" (the five listed parsers should always be installed)
-  ensure_installed = { "javascript", "html", "typescript", "markdown", "markdown_inline", "go", "rust", "c", "lua",
-    "vim", "vimdoc", "query" },
-  -- Install parsers synchronously (only applied to `ensure_installed`)
-  sync_install = false,
-  -- Automatically install missing parsers when entering buffer
-  -- Recommendation: set to false if you don't have `tree-sitter` CLI installed locally
-  auto_install = true,
-  highlight = {
-    enable = true,
-
-    -- Setting this to true will run `:h syntax` and tree-sitter at the same time.
-    -- Set this to `true` if you depend on 'syntax' being enabled (like for indentation).
-    -- Using this option may slow down your editor, and you may see some duplicate highlights.
-    -- Instead of true it can also be a list of languages
-    additional_vim_regex_highlighting = false,
-  },
-  ts_context_commentstring = {
-    enable = true,
-    enable_autcmd = false,
-    config = {
-      javascript = {
-        __default = '// %s',
-        jsx_element = '{/* %s */}',
-        jsx_fragment = '{/* %s */}',
-        jsx_attribute = '// %s',
-        comment = '// %s',
-      },
-      typescript = { __default = '// %s', __multiline = '/* %s */' },
-    },
-  },
-
-}
-
-vim.filetype.add({
-  extension = {
-    mdx = 'mdx'
-  }
-})
-
-require('treesitter-context').setup({})
-
-require('ts_context_commentstring').setup {
-  enable_autocmd = false,
-}
-
-require('Comment').setup {
-  pre_hook = require('ts_context_commentstring.integrations.comment_nvim').create_pre_hook(),
-}
-
-
-
--- think these lines are borked
--- local ft_to_parser = require("nvim-treesitter.parsers").filetype_to_parsername
--- ft_to_parser.mdx = "markdown"
