--- nvim/lua/scottykaye/config/plugins/unception.lua
@@ -1,6 +1,0 @@
-vim.api.nvim_create_autocmd("User", {
-	pattern = "UnceptionEditRequestReceived",
-	callback = function()
-		require("FTerm").toggle()
-	end,
-})
