--- nvim/lua/scottykaye/init.lua
@@ -1,1 +1,0 @@
-require("scottykaye.config.lazy")
